bug-reporting-agent/requirements.txt
flask>=3.0.0
sema-sdk>=0.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
